Move plugin commands into their own folder
