+++ lib/danger/commands/plugins/new_plugin.rb
