+++ lib/danger/commands/runner.rb
@@ -2,7 +2,7 @@ module Danger
   class Runner < CLAide::Command
     require 'danger/commands/init'
     require 'danger/commands/local'
-    require 'danger/commands/plugin_abstract'
+    require 'danger/commands/plugins/plugin_abstract'
 
     self.summary = 'Run the Dangerfile.'
     self.command = 'danger'
